# backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/reports/verification/{verificationId}/latest": {
      "get": {
        "operationId": "ReportsController_latest",
        "parameters": [
          {
            "name": "verificationId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Reports"
        ]
      }
    },
    "/api/v1/reports/verification/{verificationId}/versions": {
      "get": {
        "operationId": "ReportsController_versions",
        "parameters": [
          {
            "name": "verificationId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Reports"
        ]
      }
    },
    "/api/v1/reports/{id}": {
      "get": {
        "operationId": "ReportsController_get",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Reports"
        ]
      },
      "delete": {
        "operationId": "ReportsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Reports"
        ]
      }
    },
    "/api/v1/reports/{id}/visibility": {
      "patch": {
        "operationId": "ReportsController_visibility",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateReportVisibilityDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Reports"
        ]
      }
    },
    "/api/v1/reports/{id}/revoke": {
      "post": {
        "operationId": "ReportsController_revoke",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Reports"
        ]
      }
    },
    "/api/v1/reports/{id}/feedback": {
      "post": {
        "operationId": "ReportsController_feedback",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReportFeedbackDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Reports"
        ]
      }
    },
    "/api/v1/reports/{id}/export/json": {
      "get": {
        "operationId": "ReportsController_jsonExport",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Download a real public-safe JSON report export",
        "tags": [
          "Reports"
        ]
      }
    },
    "/api/v1/reports/{id}/export/pdf": {
      "get": {
        "operationId": "ReportsController_pdfExport",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Download a real public-safe PDF report export",
        "tags": [
          "Reports"
        ]
      }
    },
...
    "/api/v1/learning/recommendations/report/{reportId}": {
      "get": {
        "operationId": "LearningController_recommendations",
        "parameters": [
          {
            "name": "reportId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Learning"
        ]
      }
    },
...
  },
  "info": {
    "title": "Verith API",
    "description": "Explainable misinformation verification and media literacy API",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "verith_refresh": {
        "type": "apiKey",
        "in": "cookie",
        "name": "verith_refresh"
      }
    },
    "schemas": {
...
      "UpdateReportVisibilityDto": {
        "type": "object",
        "properties": {
          "visibility": {
            "type": "string",
            "enum": [
              "PRIVATE",
              "UNLISTED",
              "PUBLIC"
            ]
          }
        },
        "required": [
          "visibility"
        ]
      },
      "ReportFeedbackDto": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "enum": [
              "HELPFUL",
              "NOT_HELPFUL",
              "PROBLEM_REPORTED"
            ]
          },
          "category": {
            "type": "string",
            "enum": [
              "INCORRECT_VERDICT",
              "MISSING_CONTEXT",
              "BROKEN_SOURCE",
              "UNSAFE_CONTENT",
              "PRIVACY_CONCERN",
              "OTHER"
            ]
          },
          "comment": {
            "type": "string"
          }
        },
        "required": [
          "type"
        ]
      },
...
    }
  }
}